+++ src/main/java/roomescape/controller/ReservationController.java
@@ -1,0 +1,63 @@
+package roomescape.controller;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import roomescape.domain.Reservation;
+import roomescape.dto.ReservationRequest;
+import roomescape.exception.NotFoundReservationException;
+
+import java.net.URI;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Objects;
+import java.util.concurrent.atomic.AtomicLong;
+
+@Controller
+public class ReservationController {
+    private final List<Reservation> reservations = new ArrayList<>();
+    private final AtomicLong index = new AtomicLong(1);
+
+    /*ReservationController() {
+        reservations.add(new Reservation(1L, "브라운", "2023-01-01", "10:00"));
+        reservations.add(new Reservation(2L, "브라운", "2023-01-02", "11:00"));
+        reservations.add(new Reservation(3L, "브라운", "2023-01-03", "13:00"));
+    }*/
+
+    @PostMapping("/reservations")
+    public ResponseEntity<Reservation> create(@RequestBody ReservationRequest request) {
+        Reservation newReservation = Reservation.create(index.getAndIncrement(), request.name(), request.date(), request.time());
+        reservations.add(newReservation);
+        return ResponseEntity.created(URI.create("/reservations/" + newReservation.getId())).
+                body(newReservation);
+    }
+
+    @GetMapping("/reservation")
+    public String reservation() {
+        return "reservation";
+    }
+
+    @GetMapping("/reservations")
+    public ResponseEntity<List<Reservation>> read() {
+        return ResponseEntity
+                .ok()
+                .body(reservations);
+    }
+
+    @DeleteMapping("/reservations/{id}")
+    public ResponseEntity<Void> delete(@PathVariable Long id) {
+        Reservation reservation = reservations.stream()
+                .filter(it -> Objects.equals(it.getId(), id))
+                .findFirst()
+                .orElseThrow(() -> new NotFoundReservationException("예약된 기록이 없습니다."));
+
+        reservations.remove(reservation);
+
+        return ResponseEntity.noContent().build();
+
+    }
+}
